refactor: extract RadioStation model
